Refactor SSVC decision point handling for reuse and clarity
Moved SSVC decision point initialization to a helper module for better modularity and reuse. Simplified `test_6_1_48` logic by leveraging the centralized `CSAF_SSVC_DECISION_POINTS` and `DP_VAL_LOOKUP` structures.
